% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{consider}
\alias{consider}
\title{Car brand consideration}
\format{
A data frame with 1000 rows and 2 variables
}
\usage{
data(consider)
}
\description{
Car brand consideration
}
\details{
Survey data of consumer purchase intentions. Description provided in attr(consider,"description")
}
\keyword{datasets}
