% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{cor2df}
\alias{cor2df}
\title{Store a correlation matrix as a (long) data.frame}
\usage{
cor2df(object, labels = c("label1", "label2"), ...)
}
\arguments{
\item{object}{Return value from \code{\link{correlation}}}

\item{labels}{Column names for the correlation pairs}

\item{...}{further arguments passed to or from other methods}
}
\description{
Store a correlation matrix as a (long) data.frame
}
\details{
Return the correlation matrix as a (long) data.frame. See \url{https://radiant-rstats.github.io/docs/basics/correlation.html} for an example in Radiant
}
