% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{demand_uk}
\alias{demand_uk}
\title{Demand in the UK}
\format{A data frame with 1000 rows and 2 variables}
\usage{
data(demand_uk)
}
\description{
Demand in the UK
}
\details{
Survey data of consumer purchase intentions. Description provided in attr(demand_uk,"description")
}
\keyword{datasets}
