% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{make_vec}
\alias{make_vec}
\title{Convert a string of numbers into a vector}
\usage{
make_vec(x, char = FALSE)
}
\arguments{
\item{x}{A string of numbers that may include fractions}

\item{char}{If string contains fractions, return as a character vector}
}
\description{
Convert a string of numbers into a vector
}
\examples{
make_vec("1 2 4")
make_vec("1/2 2/3 4/5")
make_vec("1/2 2/3 4/5", char = TRUE)
make_vec("1/2 2/3 1", char = TRUE)
}
