% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_props.R
\name{plot.compare_props}
\alias{plot.compare_props}
\title{Plot method for the compare_props function}
\usage{
\method{plot}{compare_props}(x, plots = "bar", shiny = FALSE,
  custom = FALSE, ...)
}
\arguments{
\item{x}{Return value from \code{\link{compare_props}}}

\item{plots}{One or more plots of proportions ("bar" or "dodge")}

\item{shiny}{Did the function call originate inside a shiny app}

\item{custom}{Logical (TRUE, FALSE) to indicate if ggplot object (or list of ggplot objects) should be returned. This option can be used to customize plots (e.g., add a title, change x and y labels, etc.). See examples and \url{http://docs.ggplot2.org} for options.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the compare_props function
}
\details{
See \url{https://radiant-rstats.github.io/docs/basics/compare_props.html} for an example in Radiant
}
\examples{
result <- compare_props(titanic, "pclass", "survived")
plot(result, plots = c("bar", "dodge"))

}
\seealso{
\code{\link{compare_props}} to calculate results

\code{\link{summary.compare_props}} to summarize results
}
