% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{plot.prob_chisq}
\alias{plot.prob_chisq}
\title{Plot method for the probability calculator (Chi-squared distribution)}
\usage{
\method{plot}{prob_chisq}(x, type = "values", ...)
}
\arguments{
\item{x}{Return value from \code{\link{prob_chisq}}}

\item{type}{Probabilities ("probs") or values ("values")}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the probability calculator (Chi-squared distribution)
}
\details{
See \url{https://radiant-rstats.github.io/docs/basics/prob_calc.html} for an example in Radiant
}
\examples{
result <- prob_chisq(df = 1, ub = 3.841)
plot(result, type = "values")

}
\seealso{
\code{\link{prob_chisq}} to calculate results

\code{\link{summary.prob_chisq}} to summarize results
}
