% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{plot.prob_disc}
\alias{plot.prob_disc}
\title{Plot method for the probability calculator (discrete)}
\usage{
\method{plot}{prob_disc}(x, type = "values", ...)
}
\arguments{
\item{x}{Return value from \code{\link{prob_disc}}}

\item{type}{Probabilities ("probs") or values ("values")}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the probability calculator (discrete)
}
\details{
See \url{https://radiant-rstats.github.io/docs/basics/prob_calc.html} for an example in Radiant
}
\examples{
result <- prob_disc(v = 1:6, p = c(2 / 6, 2 / 6, 1 / 12, 1 / 12, 1 / 12, 1 / 12), pub = 0.95)
plot(result, type = "probs")

}
\seealso{
\code{\link{prob_disc}} to calculate results

\code{\link{summary.prob_disc}} to summarize results
}
