% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{plot.prob_lnorm}
\alias{plot.prob_lnorm}
\title{Plot method for the probability calculator (log normal)}
\usage{
\method{plot}{prob_lnorm}(x, type = "values", ...)
}
\arguments{
\item{x}{Return value from \code{\link{prob_norm}}}

\item{type}{Probabilities ("probs") or values ("values")}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the probability calculator (log normal)
}
\details{
See \url{https://radiant-rstats.github.io/docs/basics/prob_calc.html} for an example in Radiant
}
\examples{
result <- prob_lnorm(meanlog = 0, sdlog = 1, lb = 0, ub = 1)
plot(result, type = "values")

}
\seealso{
\code{\link{prob_lnorm}} to calculate results

\code{\link{plot.prob_lnorm}} to plot results
}
