% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{plot.prob_norm}
\alias{plot.prob_norm}
\title{Plot method for the probability calculator (normal)}
\usage{
\method{plot}{prob_norm}(x, type = "values", ...)
}
\arguments{
\item{x}{Return value from \code{\link{prob_norm}}}

\item{type}{Probabilities ("probs") or values ("values")}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the probability calculator (normal)
}
\details{
See \url{https://radiant-rstats.github.io/docs/basics/prob_calc.html} for an example in Radiant
}
\examples{
result <- prob_norm(mean = 0, stdev = 1, ub = 0)
plot(result)

}
\seealso{
\code{\link{prob_norm}} to calculate results

\code{\link{summary.prob_norm}} to summarize results
}
