% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{plot.prob_tdist}
\alias{plot.prob_tdist}
\title{Plot method for the probability calculator (t-distribution)}
\usage{
\method{plot}{prob_tdist}(x, type = "values", ...)
}
\arguments{
\item{x}{Return value from \code{\link{prob_tdist}}}

\item{type}{Probabilities ("probs") or values ("values")}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the probability calculator (t-distribution)
}
\details{
See \url{https://radiant-rstats.github.io/docs/basics/prob_calc.html} for an example in Radiant
}
\examples{
result <- prob_tdist(df = 10, ub = 2.228)
plot(result, type = "values")

}
\seealso{
\code{\link{prob_tdist}} to calculate results

\code{\link{summary.prob_tdist}} to summarize results
}
