% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{print.rcorr}
\alias{print.rcorr}
\title{Print method for the correlation function}
\usage{
\method{print}{rcorr}(x, ...)
}
\arguments{
\item{x}{Return value from \code{\link{correlation}}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print method for the correlation function
}
