% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{prob_binom}
\alias{prob_binom}
\title{Probability calculator for the binomial distribution}
\usage{
prob_binom(n, p, lb = NA, ub = NA, plb = NA, pub = NA, dec = 3)
}
\arguments{
\item{n}{Number of trials}

\item{p}{Probability}

\item{lb}{Lower bound on the number of successes}

\item{ub}{Upper bound on the number of successes}

\item{plb}{Lower probability bound}

\item{pub}{Upper probability bound}

\item{dec}{Number of decimals to show}
}
\description{
Probability calculator for the binomial distribution
}
\details{
See \url{https://radiant-rstats.github.io/docs/basics/prob_calc.html} for an example in Radiant
}
\examples{
prob_binom(n = 10, p = 0.3, ub = 3)

}
\seealso{
\code{\link{summary.prob_binom}} to summarize results

\code{\link{plot.prob_binom}} to plot results
}
