% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{prob_disc}
\alias{prob_disc}
\title{Probability calculator for a discrete distribution}
\usage{
prob_disc(v, p, lb = NA, ub = NA, plb = NA, pub = NA, dec = 3)
}
\arguments{
\item{v}{Values}

\item{p}{Probabilities}

\item{lb}{Lower bound on the number of successes}

\item{ub}{Upper bound on the number of successes}

\item{plb}{Lower probability bound}

\item{pub}{Upper probability bound}

\item{dec}{Number of decimals to show}
}
\description{
Probability calculator for a discrete distribution
}
\details{
See \url{https://radiant-rstats.github.io/docs/basics/prob_calc.html} for an example in Radiant
}
\examples{
prob_disc(v = "5 6 7 8 9 10 11", p = ".1 .2 .3 .15 .1 .1 .05", pub = 0.95)

}
