% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{prob_lnorm}
\alias{prob_lnorm}
\title{Probability calculator for the log normal distribution}
\usage{
prob_lnorm(meanlog, sdlog, lb = NA, ub = NA, plb = NA, pub = NA, dec = 3)
}
\arguments{
\item{meanlog}{Mean of the distribution on the log scale}

\item{sdlog}{Standard deviation of the distribution on the log scale}

\item{lb}{Lower bound (default is -Inf)}

\item{ub}{Upper bound (default is Inf)}

\item{plb}{Lower probability bound}

\item{pub}{Upper probability bound}

\item{dec}{Number of decimals to show}
}
\description{
Probability calculator for the log normal distribution
}
\details{
See \url{https://radiant-rstats.github.io/docs/basics/prob_calc.html} for an example in Radiant
}
\examples{
prob_lnorm(meanlog = 0, sdlog = 1, lb = 0, ub = 1)

}
\seealso{
\code{\link{summary.prob_lnorm}} to summarize results

\code{\link{plot.prob_lnorm}} to plot results
}
