% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{prob_tdist}
\alias{prob_tdist}
\title{Probability calculator for the t-distribution}
\usage{
prob_tdist(df, lb = NA, ub = NA, plb = NA, pub = NA, dec = 3)
}
\arguments{
\item{df}{Degrees of freedom}

\item{lb}{Lower bound (default is -Inf)}

\item{ub}{Upper bound (default is Inf)}

\item{plb}{Lower probability bound}

\item{pub}{Upper probability bound}

\item{dec}{Number of decimals to show}
}
\description{
Probability calculator for the t-distribution
}
\details{
See \url{https://radiant-rstats.github.io/docs/basics/prob_calc.html} for an example in Radiant
}
\examples{
prob_tdist(df = 10, ub = 2.228)

}
