% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{prob_unif}
\alias{prob_unif}
\title{Probability calculator for the uniform distribution}
\usage{
prob_unif(min, max, lb = NA, ub = NA, plb = NA, pub = NA, dec = 3)
}
\arguments{
\item{min}{Minimum value}

\item{max}{Maximum value}

\item{lb}{Lower bound (default = 0)}

\item{ub}{Upper bound (default = 1)}

\item{plb}{Lower probability bound}

\item{pub}{Upper probability bound}

\item{dec}{Number of decimals to show}
}
\description{
Probability calculator for the uniform distribution
}
\details{
See \url{https://radiant-rstats.github.io/docs/basics/prob_calc.html} for an example in Radiant
}
\examples{
prob_unif(min = 0, max = 1, ub = 0.3)

}
\seealso{
\code{\link{summary.prob_unif}} to summarize results

\code{\link{plot.prob_unif}} to plot results
}
