% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/radiant.R
\docType{package}
\name{radiant.basics}
\alias{radiant.basics}
\alias{radiant.basics-package}
\alias{radiant.basics}
\title{radiant.basics}
\usage{
radiant.basics(run = "browser")
}
\arguments{
\item{run}{Run radiant.basics in an external browser ("browser") or in the Rstudio viewer ("viewer")}
}
\description{
radiant.basics

Launch radiant.basics in default browser or Rstudio Viewer
}
\details{
See \url{https://radiant-rstats.github.io/docs} for documentation and tutorials
}
\examples{
\dontrun{
radiant.basics::radiant.basics()
radiant.basics::radiant.basics("viewer")
}

}
