% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{radiant.basics_viewer}
\alias{radiant.basics_viewer}
\title{Launch radiant.basics in the Rstudio viewer}
\usage{
radiant.basics_viewer()
}
\description{
Launch radiant.basics in the Rstudio viewer
}
\details{
See \url{https://radiant-rstats.github.io/docs} for documentation and tutorials
}
\examples{
\dontrun{
radiant.basics_viewer()
}
}
