% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{radiant.basics_window}
\alias{radiant.basics_window}
\title{Launch radiant.basics in an Rstudio window}
\usage{
radiant.basics_window()
}
\description{
Launch radiant.basics in an Rstudio window
}
\details{
See \url{https://radiant-rstats.github.io/docs} for documentation and tutorials
}
\examples{
\dontrun{
radiant.basics_window()
}
}
