% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{salary}
\alias{salary}
\title{Salaries for Professors}
\format{
A data frame with 397 rows and 6 variables
}
\usage{
data(salary)
}
\description{
Salaries for Professors
}
\details{
2008-2009 nine-month salary for professors in a college in the US. Description provided in attr(salary,description")
}
\keyword{datasets}
