% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{summary.prob_binom}
\alias{summary.prob_binom}
\title{Summary method for the probability calculator (binomial)}
\usage{
\method{summary}{prob_binom}(object, type = "values", ...)
}
\arguments{
\item{object}{Return value from \code{\link{prob_binom}}}

\item{type}{Probabilities ("probs") or values ("values")}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the probability calculator (binomial)
}
\details{
See \url{https://radiant-rstats.github.io/docs/basics/prob_calc.html} for an example in Radiant
}
\examples{
result <- prob_binom(n = 10, p = 0.3, ub = 3)
summary(result, type = "values")

}
