% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{summary.prob_chisq}
\alias{summary.prob_chisq}
\title{Summary method for the probability calculator function (Chi-squared distribution)}
\usage{
\method{summary}{prob_chisq}(object, type = "values", ...)
}
\arguments{
\item{object}{Return value from \code{\link{prob_chisq}}}

\item{type}{Probabilities or values}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the probability calculator function (Chi-squared distribution)
}
\details{
See \url{https://radiant-rstats.github.io/docs/basics/prob_calc.html} for an example in Radiant
}

