% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{summary.prob_fdist}
\alias{summary.prob_fdist}
\title{Summary method for the probability calculator (F-distribution)}
\usage{
\method{summary}{prob_fdist}(object, type = "values", ...)
}
\arguments{
\item{object}{Return value from \code{\link{prob_fdist}}}

\item{type}{Probabilities ("probs") or values ("values")}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the probability calculator (F-distribution)
}
\details{
See \url{https://radiant-rstats.github.io/docs/basics/prob_calc.html} for an example in Radiant
}
\examples{
result <- prob_fdist(df1 = 10, df2 = 10, ub = 2.978)
summary(result, type = "values")

}
\seealso{
\code{\link{prob_fdist}} to calculate results

\code{\link{plot.prob_fdist}} to plot results
}
