% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{summary.prob_lnorm}
\alias{summary.prob_lnorm}
\title{Summary method for the probability calculator (log normal)}
\usage{
\method{summary}{prob_lnorm}(object, type = "values", ...)
}
\arguments{
\item{object}{Return value from \code{\link{prob_norm}}}

\item{type}{Probabilities ("probs") or values ("values")}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the probability calculator (log normal)
}
\details{
See \url{https://radiant-rstats.github.io/docs/basics/prob_calc.html} for an example in Radiant
}
\examples{
result <- prob_lnorm(meanlog = 0, sdlog = 1, lb = 0, ub = 1)
summary(result, type = "values")

}
\seealso{
\code{\link{prob_lnorm}} to calculate results

\code{\link{plot.prob_lnorm}} to summarize results
}
