% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{summary.prob_pois}
\alias{summary.prob_pois}
\title{Summary method for the probability calculator (poisson)}
\usage{
\method{summary}{prob_pois}(object, type = "values", ...)
}
\arguments{
\item{object}{Return value from \code{\link{prob_pois}}}

\item{type}{Probabilities ("probs") or values ("values")}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the probability calculator (poisson)
}
\details{
See \url{https://radiant-rstats.github.io/docs/basics/prob_calc.html} for an example in Radiant
}
\examples{
result <- prob_pois(lambda = 1, ub = 3)
summary(result, type = "values")

}
