% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{summary.prob_tdist}
\alias{summary.prob_tdist}
\title{Summary method for the probability calculator (t-distribution)}
\usage{
\method{summary}{prob_tdist}(object, type = "values", ...)
}
\arguments{
\item{object}{Return value from \code{\link{prob_tdist}}}

\item{type}{Probabilities ("probs") or values ("values")}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the probability calculator (t-distribution)
}
\details{
See \url{https://radiant-rstats.github.io/docs/basics/prob_calc.html} for an example in Radiant
}
\examples{
result <- prob_tdist(df = 10, ub = 2.228)
summary(result, type = "values")

}
