% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{summary.prob_unif}
\alias{summary.prob_unif}
\title{Summary method for the probability calculator (uniform)}
\usage{
\method{summary}{prob_unif}(object, type = "values", ...)
}
\arguments{
\item{object}{Return value from \code{\link{prob_unif}}}

\item{type}{Probabilities ("probs") or values ("values")}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the probability calculator (uniform)
}
\details{
See \url{https://radiant-rstats.github.io/docs/basics/prob_calc.html} for an example in Radiant
}
\examples{
result <- prob_unif(min = 0, max = 1, ub = 0.3)
summary(result, type = "values")

}
