% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_prop.R
\name{summary.single_prop}
\alias{summary.single_prop}
\title{Summary method for the single_prop function}
\usage{
\method{summary}{single_prop}(object, dec = 3, ...)
}
\arguments{
\item{object}{Return value from \code{\link{single_prop}}}

\item{dec}{Number of decimals to show}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the single_prop function
}
\details{
See \url{https://radiant-rstats.github.io/docs/basics/single_prop.html} for an example in Radiant
}
\examples{
result <- single_prop(titanic, "survived", lev = "Yes", comp_value = 0.5, alternative = "less")
summary(result)

}
\seealso{
\code{\link{single_prop}} to generate the results

\code{\link{plot.single_prop}} to plot the results
}
