% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clt.R
\name{plot.clt}
\alias{plot.clt}
\title{Plot method for the Central Limit Theorem simulation}
\usage{
\method{plot}{clt}(x, stat = "sum", bins = 15, ...)
}
\arguments{
\item{x}{Return value from \code{\link{clt}}}

\item{stat}{Statistic to use (sum or mean)}

\item{bins}{Number of bins to use}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the Central Limit Theorem simulation
}
\details{
See \url{https://radiant-rstats.github.io/docs/basics/clt.html} for an example in Radiant
}
\examples{
clt("Uniform", 100, 100, unif_min = 10, unif_max = 20) \%>\% plot()

}
