% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{plot.prob_pois}
\alias{plot.prob_pois}
\title{Plot method for the probability calculator function (Poisson distribution)}
\usage{
\method{plot}{prob_pois}(x, type = "values", ...)
}
\arguments{
\item{x}{Return value from \code{\link{prob_pois}}}

\item{type}{Probabilities or values}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the probability calculator function (Poisson distribution)
}
\details{
See \url{https://radiant-rstats.github.io/docs/basics/prob_calc.html} for an example in Radiant
}
