% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{plot.prob_unif}
\alias{plot.prob_unif}
\title{Plot method for the probability calculator (uniform)}
\usage{
\method{plot}{prob_unif}(x, type = "values", ...)
}
\arguments{
\item{x}{Return value from \code{\link{prob_unif}}}

\item{type}{Probabilities ("probs") or values ("values")}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot method for the probability calculator (uniform)
}
\details{
See \url{https://radiant-rstats.github.io/docs/basics/prob_calc.html} for an example in Radiant
}
\examples{
result <- prob_unif(min = 0, max = 1, ub = 0.3)
plot(result, type = "values")

}
