% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{prob_pois}
\alias{prob_pois}
\title{Probability calculator for the poisson distribution}
\usage{
prob_pois(lambda, lb = NA, ub = NA, plb = NA, pub = NA, dec = 3)
}
\arguments{
\item{lambda}{Rate}

\item{lb}{Lower bound (default is 0)}

\item{ub}{Upper bound (default is Inf)}

\item{plb}{Lower probability bound}

\item{pub}{Upper probability bound}

\item{dec}{Number of decimals to show}
}
\description{
Probability calculator for the poisson distribution
}
\details{
See \url{https://radiant-rstats.github.io/docs/basics/prob_calc.html} for an example in Radiant
}
\examples{
prob_pois(lambda = 1, ub = 3)

}
\seealso{
\code{\link{summary.prob_pois}} to summarize results

\code{\link{plot.prob_pois}} to plot results
}
