% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/radiant.R
\docType{package}
\name{radiant.basics}
\alias{radiant.basics}
\alias{radiant.basics-package}
\title{radiant.basics}
\usage{
radiant.basics()
}
\description{
radiant.basics

Launch Radiant in the default browser
}
\details{
See \url{https://radiant-rstats.github.io/docs} for documentation and tutorials
}

