% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_calc.R
\name{summary.prob_disc}
\alias{summary.prob_disc}
\title{Summary method for the probability calculator (discrete)}
\usage{
\method{summary}{prob_disc}(object, type = "values", ...)
}
\arguments{
\item{object}{Return value from \code{\link{prob_disc}}}

\item{type}{Probabilities ("probs") or values ("values")}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the probability calculator (discrete)
}
\details{
See \url{https://radiant-rstats.github.io/docs/basics/prob_calc.html} for an example in Radiant
}
\examples{
result <- prob_disc(v = 1:6, p = c(2 / 6, 2 / 6, 1 / 12, 1 / 12, 1 / 12, 1 / 12), pub = 0.95)
summary(result, type = "probs")

}
\seealso{
\code{\link{prob_disc}} to calculate results

\code{\link{plot.prob_disc}} to plot results
}
